Fix writing I2C messages when I2C addr >= 0x80
Since strings in python3 are UTF-8, calling chr() on an I2C address
>= 0x80 would create an invalid UTF-8 character (0x80 to 0xFF are
illegal). Then later, when encode() was called in add_string(), a
UnicodeEncodeError exception would be thrown because of this.

This adds a new add_bytes() method to the Telegram class for adding
bytes objects. Then, this new method is used in ls_write() so that
we can pass arbitrary I2C addresses without error.