Completely rewrite nxt.locator
nxt.locator.find_one_brick() has been removed and is replaced with
nxt.locator.find().

The new locator is simpler an relies on the backends to do backend
specific things. It also provides a function to be used in scripts to
choose the brick from the command line, and a new flexible configuration
file.

The result should be hopefully a smoother user experience.