Adds support for HiTechnic SuperPro breakout
This commit adds support for the HiTechnic SuperPro breadboardable
breakout. It's no longer sold and the documentation is hard to
find, but I have one of them and tested most of the functionality.

The only functionality that doesn't make sense to me is the
"Strobe" functionality; it appears to be simply digital outputs,
but there is not much documentation about what it does. Nothing
interesting shows up on the o-scope while using this part of the
device.

Also includes a demo of a bar graph doing a binary count from 0 to
255. Looked at examples provided in PDF but some depend on board
giving fast response time so decided against it. Should be
relatively performant.