Add specific errors subclasses for message_read
When reading a message queue, it is useful to make the distinction
between several possible errors. Add DirectProtocolError exception
subclasses: NoActiveProgramError and EmptyMailboxError.

Also update stop_program and get_current_program_name documentation to
reflect the change.

This is backward compatible as they are subclasses of the original error
exceptions.